perbaikan bug ceklis report migrasi mesin

+++ app/Database/Migrations/2026-07-20-030711_AddLineToApprovalBulanan.php
@@ -20,6 +20,8 @@ class AddLineToApprovalBulanan extends Migration
 
     public function down()
     {
-        $this->forge->dropColumn('approval_bulanan', 'line');
+        if ($this->forge->getConnection()->fieldExists('line', 'approval_bulanan')) {
+            $this->forge->dropColumn('approval_bulanan', 'line');
+        }
     }
 }
